update bonus reward calculation for fiso reward endpoint

+++ src/TeddySwap.Common/Models/Response/FisoRewardBreakdownResponse.cs
@@ -4,6 +4,6 @@ public class FisoRewardBreakdownResponse
 {
     public string StakeAddress { get; init; } = string.Empty;
     public List<FisoRewardResponse> FisoRewards { get; init; } = new();
-    public ulong TotalBaseReward { get; init; }
-    public ulong TotalBonusReward { get; init; }
+    public double TotalBaseReward { get; init; }
+    public double TotalBonusReward { get; init; }
 }
