+++ src/TeddySwap.Common/Models/Response/FisoRewardResponse.cs
@@ -5,6 +5,6 @@ public class FisoRewardResponse
     public string PoolId { get; init; } = string.Empty;
     public ulong Epoch { get; init; }
     public ulong ActiveStake { get; init; }
-    public ulong BaseReward { get; init; }
-    public ulong BonusReward { get; set; }
+    public double BaseReward { get; init; }
+    public double BonusReward { get; set; }
 }
